backend/requirements.txt
Brotli==1.1.0
CacheControl==0.14.3
Crawl4AI==0.6.3
Flask-Compress==1.14
Flask-Cors==4.0.1
Flask==3.0.3
Jinja2==3.1.6
Levenshtein==0.27.1
MarkupSafe==3.0.2
PyJWT==2.10.1
PySocks==1.7.1
PyYAML==6.0.2
Pygments==2.19.1
RapidFuzz==3.13.0
Werkzeug==3.1.3
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
cerebras_cloud_sdk==1.35.0
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.4
cssselect==1.3.0
distro==1.9.0
docker==7.0.0
fake-http-header==0.3.5
fake-useragent==2.2.0
filelock==3.18.0
firebase-admin==6.9.0
frozenlist==1.7.0
fsspec==2025.5.1
fuzzywuzzy==0.18.0
git-filter-repo==2.47.0
kubernetes==31.0.0
google-api-core==2.25.1
google-api-python-client==2.174.0
google-auth-httplib2==0.2.0
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.1.1
google-crc32c==1.7.1
google-genai==1.23.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.3
groq==0.13.0
grpcio-status==1.73.1
grpcio==1.73.1
gunicorn==21.2.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.33.0
humanize==4.12.3
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
jiter==0.10.0
joblib==1.5.1
jsonschema-specifications==2025.4.1
jsonschema==4.24.0
litellm==1.72.4
lxml==5.4.0
markdown-it-py==3.0.0
mdurl==0.1.2
msgpack==1.1.1
multidict==6.4.4
nltk==3.9.1
numpy>=1.26.0
openai==1.86.0
outcome==1.3.0.post0
packaging==25.0
pandas>=2.0.0
pillow==10.4.0
playwright==1.52.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
pyOpenSSL==25.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pyee==13.0.0
pyparsing==3.2.3
pyperclip==1.9.0
pytest==8.3.4
python-Levenshtein==0.27.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
rank-bm25==0.2.2
referencing==0.36.2
regex==2024.11.6
requests<2.32.0
rich==13.9.4
rpds-py==0.25.1
rsa==4.9.1
selenium==4.34.0
six==1.17.0
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.7
tenacity==8.5.0
tf-playwright-stealth==1.1.2
tiktoken==0.9.0
tokenizers==0.21.1
tqdm==4.67.1
trio-websocket==0.12.2
trio==0.30.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
upstash-redis==1.1.0
uritemplate==4.2.0
urllib3==2.4.0
watchdog==6.0.0
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
Flask-Limiter==3.8.0
Flask-Talisman==1.1.0
opensearch-py==2.4.2
Pillow
gevent>=24.2.1
